Read two signed quantities as the signed numbers they are
`Bitvector32Term::as_const` answers `u32`, and two rules that read a signed
`int32` quantity through it then ordered the result. `e219bbc8` is the same
bug on range endpoints; these two are on a block extent and a resource
quantity.

**A view below its block.** A caller lends a view over its own stack object
without a resource clause, because the object is its own, and
`local_view_range_within_block` is what says the block has the storage. It
took `end <= start` first, as "an empty range names no storage". A start of
`-1` arrives there as `4294967295`, so every range beginning below its block
took that exit and `memory.access_in_bounds` never ran:

    int32 first(int32* values, int32 s, int32 e) {
        return values[0];
    }                                    //  views values[s..e]

    int32 lends_below_its_local() {
        int32 pair[2];
        pair[0] = 1;
        pair[1] = 2;
        return first(pair, -1, 3);       //  lends pair[-1..3]
    }

`pair` is two elements and the call lent four, one below the object and one
past it. The contained range is the control: the same call at `0, 3` is
refused with `missing resource fact`, so a *superset* of a refused range was
admitted. The callee can reach the out-of-block half — the same shape with
`return values[2]` verifies the callee and stops in the caller only at the
uninitialized-storage backstop, one check later than the contained range
stops.

Both endpoints are now read signed, which is the reading
`memory_range_byte_count` already scales, and the byte count is formed in
`i64` before it is narrowed. The `end <= start` exit stays: a backwards range
is how a range says it is empty.

**A negative fold quantity.** `fold(n of R(a))` reads `n` twice — once for
"proved zero or positive", once to decide whether the body is active and so
consumed — and the constant arm asked `as_const`. `-1` was positive by that
reading, so `fold(-1 of ref(o))` verified and left the body active. The
symbolic arm beside it always asked `Bitvector32SignedGreaterThan`, so the
two arms disagreed about one number. The population it leaves really counts
`-1`: stating `ensures count(ref(o)) == 0` afterwards reports "left side
evaluated to -1".

Neither is spent in the corpus — `resource_quantity_at_least` is signed, so
`owns -1 of R` entails no positive count, and the declared-quantity gate in
`evaluate_function_resource_spec` asks the signed question — so both are
accepted steps rather than false `ensures`. An accepted step that admits a
ghost state no nonnegativity invariant can describe is the hole, and the
constant arm agreeing with its own symbolic arm is the fix.

Tests: `mdtests/a_local_view_below_its_block_is_not_lent.md` and
`mdtests/fold_rejects_a_negative_quantity.md`, each with its positive next
door — the caller lending the two elements it has, and the zero quantity the
gate admits with the body inactive.

+++ src/surface/proof/resources.rs
@@ -3531,7 +3531,12 @@ fn fold_composite_resources_on_outcome_with_facts(
                     ),
                     true,
                 ));
-            let quantity_is_positive = quantity.as_const().is_some_and(|value| value > 0)
+            // A resource quantity is a signed `int32`, and `as_const` answers
+            // `u32`, so `-1` read here as `4294967295` was both "positive"
+            // and the thing that made the body active. The symbolic arm
+            // beside it always asked the signed question; reading the
+            // constant signed is what makes the two agree.
+            let quantity_is_positive = quantity.as_const().is_some_and(|value| (value as i32) > 0)
                 || assumptions.proves(&Proposition::ConditionIs(
                     ConditionTerm::Bitvector32SignedGreaterThan(
                         Box::new(quantity.clone()),
